moved all common functions to common.go file
