+++ handlers/common.go
@@ -3,7 +3,11 @@ package handlers
 import (
 	"context"
 
+	"net/http"
+
 	"github.com/ONSdigital/dp-api-clients-go/v2/dataset"
+
+	"github.com/ONSdigital/log.go/v2/log"
 )
 
 func getReleaseDate(ctx context.Context, dc DatasetClient, userAuthToken, collectionID, datasetID, edition, versionID string) (string, error) {
@@ -24,4 +28,13 @@ func getReleaseDate(ctx context.Context, dc DatasetClient, userAuthToken, collec
 	} else {
 		return version.ReleaseDate, nil
 	}
+}
+
+func setStatusCode(req *http.Request, w http.ResponseWriter, err error) {
+	status := http.StatusInternalServerError
+	if err, ok := err.(ClientError); ok {
+		status = err.Code()
+	}
+	log.Error(req.Context(), "setting-response-status", err)
+	w.WriteHeader(status)
 }
